fix: issues with two remaining themes.
Both one-dark and nocturne used global-doc-comment instead of the correctly named doc-global-comment. Typos in Nocturned didn't prevent the theme from loading, but left annoying messages in the log. As explained by the author, the base colormap must be one of the three that imhex loads by default, dark,light and/or classic.